fix(android): avoid storing version code as signed long long
Version code (on config) is explicitly an integer, which would only
parse the initial bits of the longer value (when running under 32-bit),
resulting in a 0 value, which was then ignored as 0 is not a valid
version code.

This change makes version code a 32-bit int everywhere (including in the
App interface, which didn't have this problem because VersionCode on App
is a Number and parsed from a long long correctly).

This also fixes an undetected issue in parsing line number in event
callbacks.